+++ src/Slider.d.ts
@@ -17,7 +17,7 @@ declare class Slider extends Vue {
   classes?: object;
   showTooltip?: 'always'|'focus'|'drag';
   options?: object;
-  tooltipDirection?: null|'top'|'bottom'|'left'|'right';
+  tooltipPosition?: null|'top'|'bottom'|'left'|'right';
 
   $emit(eventName: 'change', e: {originalEvent: Event, value: any}): this;
   $emit(eventName: 'update', e: {originalEvent: Event, value: any}): this;